refactor(visual-regions): split candidate pipeline modules
Separate raster, vector, table, annotation, merge, predicate, and suppression logic from the visual region orchestrator so the main pipeline stays readable while preserving behavior.
