+++ src/core/visualRegions/types.ts
@@ -1,4 +1,26 @@
-import type { VisualRegionAssociatedText, VisualRegionKind, VisualRegionSource } from '../../types/index.js';
+import type {
+  FormField,
+  ImageBox,
+  PageAnnotation,
+  PageLayout,
+  PageQuality,
+  VectorBox,
+  VisualRegionAssociatedText,
+  VisualRegionKind,
+  VisualRegionSource,
+} from '../../types/index.js';
+
+export interface BuildVisualRegionsInput {
+  pageWidth: number;
+  pageHeight: number;
+  imageBoxes: readonly ImageBox[];
+  vectorBoxes?: readonly VectorBox[];
+  layout?: PageLayout;
+  formFields?: readonly FormField[];
+  annotations?: readonly PageAnnotation[];
+  visualStatus?: 'ok' | 'sparse' | 'blank';
+  nativeTextStatus?: PageQuality['nativeTextStatus'];
+}
 
 export interface BoxLike {
   x: number;
